+++ src/FSharp.Formatting.Literate/ParseScript.fs
@@ -432,50 +432,56 @@ type internal ParseScript(parseOptions, ctx: CompilerContext) =
             rootInputFolder = rootInputFolder
         )
 
+    /// Parse the file (without type checking) and return the text of all block comments
+    /// in source order, including the comment delimiters.
+    static member ParseBlockComments(fileName: string) : string list =
+        let sourceText = (System.IO.File.ReadAllText >> SourceText.ofString) fileName
+        let checker = FSharp.Formatting.Internal.CompilerServiceExtensions.FSharpAssemblyHelper.checker
+
+        let parseResult =
+            checker.ParseFile(
+                fileName,
+                sourceText,
+                { FSharpParsingOptions.Default with
+                    SourceFiles = [| fileName |]
+                }
+            )
+            |> Async.RunSynchronously
+
+        match parseResult.ParseTree with
+        | ParsedInput.SigFile _ -> []
+        | ParsedInput.ImplFile(ParsedImplFileInput.ParsedImplFileInput(trivia = { CodeComments = codeComments })) ->
+            codeComments
+            |> List.choose (function
+                | CommentTrivia.BlockComment mBlockComment -> Some mBlockComment
+                | CommentTrivia.LineComment _ -> None)
+            |> List.sortBy (fun m -> m.StartLine, m.StartColumn)
+            |> List.map (fun mBlockComment ->
+                // Grab the comment text from the ISourceText
+                let startLine = mBlockComment.StartLine - 1
+                let line = sourceText.GetLineString startLine
+
+                if mBlockComment.StartLine = mBlockComment.EndLine then
+                    let length = mBlockComment.EndColumn - mBlockComment.StartColumn
+                    line.Substring(mBlockComment.StartColumn, length)
+                else
+                    let firstLineContent = line.Substring(mBlockComment.StartColumn)
+                    let sb = StringBuilder().AppendLine(firstLineContent)
+
+                    (sb, [ mBlockComment.StartLine .. mBlockComment.EndLine - 2 ])
+                    ||> List.fold (fun sb lineNumber -> sb.AppendLine(sourceText.GetLineString lineNumber))
+                    |> fun sb ->
+                        let lastLine = sourceText.GetLineString(mBlockComment.EndLine - 1)
+                        sb.Append(lastLine.Substring(0, mBlockComment.EndColumn)).ToString())
+
     /// Tries and parse the file to find and process the first block comment.
     static member ParseFrontMatter(fileName: string) : FrontMatterFile option =
         try
-            let sourceText = (System.IO.File.ReadAllText >> SourceText.ofString) fileName
-            let checker = FSharp.Formatting.Internal.CompilerServiceExtensions.FSharpAssemblyHelper.checker
-
-            let parseResult =
-                checker.ParseFile(
-                    fileName,
-                    sourceText,
-                    { FSharpParsingOptions.Default with
-                        SourceFiles = [| fileName |]
-                    }
-                )
-                |> Async.RunSynchronously
-
-            match parseResult.ParseTree with
-            | ParsedInput.SigFile _ -> None
-            | ParsedInput.ImplFile(ParsedImplFileInput.ParsedImplFileInput(trivia = { CodeComments = codeComments })) ->
-                codeComments
-                |> List.tryPick (function
-                    | CommentTrivia.BlockComment mBlockComment -> Some mBlockComment
-                    | CommentTrivia.LineComment _ -> None)
-                |> Option.bind (fun mBlockComment ->
-                    // Grab the comment text from the ISourceText
-                    let commentText =
-                        let startLine = mBlockComment.StartLine - 1
-                        let line = sourceText.GetLineString startLine
-
-                        if mBlockComment.StartLine = mBlockComment.EndLine then
-                            let length = mBlockComment.EndColumn - mBlockComment.StartColumn
-                            line.Substring(mBlockComment.StartColumn, length)
-                        else
-                            let firstLineContent = line.Substring(mBlockComment.StartColumn)
-                            let sb = StringBuilder().AppendLine(firstLineContent)
-
-                            (sb, [ mBlockComment.StartLine .. mBlockComment.EndLine - 2 ])
-                            ||> List.fold (fun sb lineNumber -> sb.AppendLine(sourceText.GetLineString lineNumber))
-                            |> fun sb ->
-                                let lastLine = sourceText.GetLineString(mBlockComment.EndLine - 1)
-                                sb.Append(lastLine.Substring(0, mBlockComment.EndColumn)).ToString()
-
-                    let lines = commentText.Split '\n'
-                    FrontMatterFile.ParseFromLines fileName lines)
+            match ParseScript.ParseBlockComments fileName with
+            | [] -> None
+            | commentText :: _ ->
+                let lines = commentText.Split '\n'
+                FrontMatterFile.ParseFromLines fileName lines
         with ex ->
             printfn "Failed to find frontmatter in %s, %A" fileName ex
             None